- cflagIsDefined na kolumnach które zostały zdefiniowane w programie (i cflagIsLoaded na załadowanych z pliku) - obsługa ctypeWideString - usunięte trochę warningów
git-svn-id: http://www.konnekt.info/svn/staminalib/trunk/Stamina@37 e7532c55-18f9-0310-997f-f942ee4ad645

+++ Time64.h
@@ -146,7 +146,7 @@ namespace Stamina {
 		}
 
 		int toDays() const {
-			return floor((float)this->sec / (60*60*24));
+			return (int)floor((float)this->sec / (60*60*24));
 		}
 
 	private:
